requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1
torchvision==0.20.1
# diffusers must be installed from source — textual_inversion.py requires >= 0.38.0.dev0
git+https://github.com/huggingface/diffusers.git
accelerate>=0.28.0
transformers>=4.38.0
huggingface_hub>=0.22.0
Pillow>=10.0.0
ftfy>=6.1.0
tensorboard>=2.15.0
safetensors>=0.4.2
requests>=2.31.0
numpy<2
